client: fix concurrency, lifecycle, and validation bugs found by roast
Adversarial review of the go-client branch surfaced 14 verified bugs across
the client engine, live buffer, and CLI. All are fixed here with regression
tests verified to fail without their fix.

Correctness / concurrency:
- downloader: widen block-range loop counter to uint64 so a range ending at
  MaxUint32 (which the planner permits) does not wrap to 0 and loop forever.
- engine: cancel the live consumer + wait when flipAndDrain returns a replay
  error, instead of falling into liveWG.Wait() and deadlocking a quiet tail.
- livesink: record a buffering-phase append failure as fatalErr and surface it
  from flipAndDrain, rather than swallowing it and replaying a truncated buffer
  (silent data loss across the cutover).
- batcher: add emitError so recoverable errors are emitted under the same mutex
  as batches (no concurrent yield with the flusher) and a rejected error stops
  batching and fires onStop. Register onStop before starting the flusher in both
  engine paths so a racing stop cannot latch the once-guard with a nil onStop.

Lifecycle:
- Client.Close now cancels a running Events via a run-scoped cancel, honoring
  the documented "natural way to stop a live tail" contract.
- fileLiveBuffer: reopen the original file if os.Rename fails during Truncate
  (was stranding a closed fd, permanently breaking the buffer); close the read
  handle before the rename (Windows replace-open-file); guard Append/Replay/
  Truncate against use-after-close with errBufferClosed instead of nil-deref.

Validation / error classification:
- planner: reject afterSeq/beforeSeq > MaxInt64 before they wrap negative onto
  the int64 lexicon fields and silently mis-plan.
- Add ErrFatal sentinel: the engine wraps terminal failures (seed, plan, broken
  cutover) so consumers can distinguish them from recoverable per-entry errors.
  cmd/client now exits non-zero on a fatal stream error in both print and
  throughput modes (was logging/dropping and exiting 0), and rejects negative
  seq cursors instead of silently treating them as unset.

Full suite green under -race; golangci-lint clean.

ROAST_SKIP=1: roast on these files already run this session (Codex finder +
Bedrock verifier), converged clean after fixes — re-running the quick-mode
gate would duplicate a review already done with verification.

+++ internal/client/downloader.go
@@ -177,7 +177,11 @@ func (d *Downloader) downloadWholeSegment(ctx context.Context, entry PlanEntry) 
 func (d *Downloader) downloadBlocks(ctx context.Context, entry PlanEntry) ([]Event, error) {
 	var events []Event
 	for _, br := range entry.Blocks {
-		for idx := br.First; idx <= br.Last; idx++ {
+		// idx is widened to uint64 so a range ending at the uint32 max
+		// (math.MaxUint32 passes the planner's `> MaxUint32` validation) does
+		// not wrap back to 0 on the final increment and loop forever. The body
+		// only runs for idx <= br.Last <= MaxUint32, so int64/int narrowing is safe.
+		for idx := uint64(br.First); idx <= uint64(br.Last); idx++ {
 			frame, err := jetstream.JetstreamGetBlock(ctx, d.xc, int64(idx), entry.SegmentName)
 			if err != nil {
 				return nil, fmt.Errorf("jetstream: getBlock %d of %q: %w", idx, entry.SegmentName, err)
